+++ src/ONTraC/utils/niche_net_constr.py
@@ -91,7 +91,9 @@ def construct_niche_network_sample(options: Values, sample_data_df: pd.DataFrame
         3. calculate cell type composition and save it
     """
 
-    n_local = 20
+    info(f'Constructing niche network for sample: {sample_name}.')
+
+    n_local = 20  # TODO: make it as an option
     N = sample_data_df.shape[0]
 
     # get coordinates
@@ -166,6 +168,8 @@ def gen_samples_yaml(options: Values, ori_data_df: pd.DataFrame) -> None:
     :param ori_data_df: pd.DataFrame, original data
     :return: None
     """
+
+    info('Generating samples.yaml file.')
 
     data: Dict[str, List[Any]] = {'Data': []}
     for sample in ori_data_df['Sample'].unique():